SqlDelight Migration

+++ app/src/main/java/jp/ikigai/cash/flow/ui/screenStates/listing/CategoryScreenState.kt
@@ -5,7 +5,7 @@ import java.util.Locale
 
 data class CategoryScreenState(
     val categories: List<CommonListingDTO> = emptyList(),
-    val count: Int = 0,
+    val count: Long = 0,
     val countString: String = "",
     val loading: Boolean = true,
     val locale: Locale? = null
