+++ app/src/main/java/jp/ikigai/cash/flow/ui/screenStates/listing/TransactionTemplateScreenState.kt
@@ -5,7 +5,7 @@ import java.util.Locale
 
 data class TransactionTemplateScreenState(
     val templates: List<TemplateWithChips> = emptyList(),
-    val count: Int = 0,
+    val count: Long = 0,
     val countString: String = "",
     val loading: Boolean = true,
     val locale: Locale? = null
